+++ craps-simulator/Models/BetResult.cs
@@ -6,6 +6,6 @@ namespace craps_simulator.Models {
         public decimal Bet { get; set; } = 0;
         public bool IsWinner { get; set; } = false;
         public bool IsLoser { get; set; } = false;
-        public decimal Pays { get; set; } = 0;
+        public int Pays { get; set; } = 0;
     }
 }
